+++ internal/forge/gitlab/auth_connector.go
@@ -3,7 +3,6 @@ package gitlab
 import (
 	"github.com/git-town/git-town/v21/internal/cli/print"
 	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
-	"github.com/git-town/git-town/v21/internal/git/gitdomain"
 	. "github.com/git-town/git-town/v21/pkg/prelude"
 	gitlab "gitlab.com/gitlab-org/api/client-go"
 )
@@ -42,17 +41,5 @@ func (self AuthConnector) VerifyConnection() forgedomain.VerifyConnectionResult 
 		AuthenticatedUser:   NewOption(user.Username),
 		AuthenticationError: nil,
 		AuthorizationError:  err,
-	}
-}
-
-func parseMergeRequest(mergeRequest *gitlab.BasicMergeRequest) forgedomain.ProposalData {
-	return forgedomain.ProposalData{
-		MergeWithAPI: true,
-		Number:       mergeRequest.IID,
-		Source:       gitdomain.NewLocalBranchName(mergeRequest.SourceBranch),
-		Target:       gitdomain.NewLocalBranchName(mergeRequest.TargetBranch),
-		Title:        mergeRequest.Title,
-		Body:         NewOption(mergeRequest.Description),
-		URL:          mergeRequest.WebURL,
 	}
 }
